doc/memory-barriers: fix missed renaming: s/lock/acquire
Terms `lock` and `unlock` have changed to `acquire` / `release` by
commit 2e4f5382d12a441b5cccfdde00308df15c2ce300 ("locking/doc: Rename
LOCK/UNLOCK to ACQUIRE/RELEASE").  However, the commit missed to change
the table of content.  This commit changes the missed parts.
Also, section name `Acquiring functions` is not appropriate for the
section because the section is saying about lock in actual.  This commit
changes the name to more appropriate name, `Lock acquisition functions`.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>
Suggested-by: David Howells <dhowells@redhat.com>