Doc/memory-barriers: add Korean translation
This commit adds Korean version of memory-barriers.txt document.  The
header is refered to HOWTO Korean version.

The translator, SeongJae Park, is researching about multi-core
performance scalability and translating[1] a book[2] about the topic that
written by the author of the original document, Paul McKenny.

The translation has started from Feb 16, 2016 and using a github public
repository[3] to maintain the work.  The original document has updated
after the translation has started but the translation is following[4]
the whole changes[4].

Because the translator has knowledge about the topic and already
following up the upstream changes, one would sure that this translation
will keep reasonable quality and fast update speed in future, too.

[1] https://git.kernel.org/cgit/linux/kernel/git/paulmck/perfbook.git/commit/FAQ.txt?id=edbfcdee0460
[2] https://www.kernel.org/pub/linux/kernel/people/paulmck/perfbook/perfbook.html
[3] https://github.com/sjp38/linux.doc_trans_membarrier
[4] https://github.com/sjp38/linux.doc_trans_membarrier/commits/follow_origin_update

Signed-off-by: SeongJae Park <sj38.park@gmail.com>
Acked-by: David Howells <dhowells@redhat.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Acked-by: Minchan Kim <minchan@kernel.org>