Doc/memory-barriers: Add Korean translation
This commit adds Korean version of memory-barriers.txt document.  The
header is refered to HOWTO Korean version.

The translator, SeongJae Park, is interested in parallel programming and
translating[1] a book[2] about the topic.

The translation has started from Feb, 2016 and using a github public
repository[3] to maintain the work.  The work is following[4] updates to
the original document as well.

Because the translator has knowledge about the topic and already
following up the upstream changes, one would sure that this translation
will keep reasonable quality and freshness.

[1] https://git.kernel.org/cgit/linux/kernel/git/paulmck/perfbook.git/commit/FAQ.txt?id=edbfcdee0460
[2] https://www.kernel.org/pub/linux/kernel/people/paulmck/perfbook/perfbook.html
[3] https://github.com/sjp38/linux.doc_trans_membarrier
[4] https://github.com/sjp38/linux.doc_trans_membarrier/commit/06bd12d390f164fd253b861fc3aa8006d6d19ed9

Signed-off-by: SeongJae Park <sj38.park@gmail.com>
Acked-by: David Howells <dhowells@redhat.com>
Signed-off-by: Parl E. McKenney <paulmck@linux.vnet.ibm.com>
Acked-by: Minchan Kim <minchan@kernel.org>